requirements.txt
requests
Pillow
rembg[cli,cpu]
gspread
google-auth
google-auth-oauthlib
pyinstaller
pyarmor
python-dotenv
